refactor/cleanup: - remove k8s package - merge k8sman and basecrawlmanager into crawlmanager - move templates, k8sapi, utils into root package - delete all *_job.py files - rename btrix_job.yaml -> crawl_job.yaml - remove dt_now, ts_now from crawls, now in utils
